Ember.A is a runtime concern metal should not use it
 * move Ember.computed.collect back to runtime (its test are already there)
